flood-detection/pyproject.toml
[project]
name = "flood-detection"
version = "1.0.0"
description = "Karachi flood detection ML service"
requires-python = ">=3.9"
dependencies = [
    "openmeteo-requests>=1.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "requests-cache>=1.1.0",
    "retry-requests>=2.0.0",
    "scikit-learn>=1.3.0",
    "python-dotenv>=1.0.0",
]
